requirements.txt
libs

requests
scrapy
pynacl
selenium
webdriver-manager